% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{select.list}
\alias{select.list}
\alias{select_.list}
\alias{rename.list}
\alias{rename_.list}
\title{Select/rename list's elements by name}
\usage{
\method{select}{list}(.list, ...)

\method{select_}{list}(.list, ..., .dots = list())

\method{rename}{list}(.list, ...)

\method{rename_}{list}(.list, ..., .dots = list())
}
\description{
`select.list()` keeps only the variables you mention; `rename.list()`
keeps all variables. See `help(select, package = "dplyr")`` for further details.
}
