% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{select.list}
\alias{select.list}
\title{Select/rename list's elements by name}
\usage{
\method{select}{list}(.data, ...)
}
\description{
`select.data()` keeps only the elements you mention; `rename.data()`
keeps all elements. The list must be named. See `help(select, package = "dplyr")`` for further details.
}
